edit feehistory
